pyproject.toml
[project]
name = "enterprise-ai-platform"
version = "0.1.0"
description = "Enterprise AI agent platform for automating ERP implementation workflows"
requires-python = ">=3.12"
dependencies = [
    # Orchestration — framework core, needed from Layer 3 onward
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.3",
    "langfuse>=2.0",             # MIT, self-hosted — replaces LangSmith (commercial BSL)
    # Schemas / validation — needed from Layer 1 onward
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    # Storage — lightweight async clients
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "redis>=5.0",
    "psycopg2-binary>=2.9",
    "psycopg[binary]>=3.1",
    "langgraph-checkpoint-postgres>=2.0",
    # API / workers
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "celery[redis]>=5.3",
    "python-multipart>=0.0.9",
    # LLM SDK (direct — also a transitive dep of langchain-anthropic)
    "anthropic>=0.40",
    # Observability
    "structlog>=24.1",
    "prometheus-client>=0.20",
    # Templating
    "jinja2>=3.1",
    # Utilities
    "click>=8.1",
    "httpx>=0.27",
    "networkx>=3.2",
    "numpy>=1.26",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Heavy ML/parsing stack — installed from Layer 2 onward (make setup installs this)
ml = [
    # Document parsing (text-only — no ML models, no OCR)
    "pypdf>=4.0",
    "python-docx>=1.1",
    "python-magic>=0.4",

    # NLP
    "spacy>=3.7",
    "rapidfuzz>=3.6",

    # Vector / retrieval (fastembed uses ONNX Runtime ~50 MB — no PyTorch)
    "qdrant-client>=1.8",
    "fastembed>=0.3",
    "rank-bm25>=0.2",
]

# Dev tooling only — what CI installs for Layer 0/1
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "types-PyYAML>=6.0",
    "types-redis>=4.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["platform", "agents", "modules", "api"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "modules/dynafit/tests"]
pythonpath = ["."]
markers = [
    "unit: fast, no external deps",
    "integration: needs Docker services",
    "golden: uses golden fixture files",
    "llm: needs live LLM (skip in CI)",
]
addopts = "--strict-markers"

# ---------------------------------------------------------------------------
# ruff
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line-length enforced by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["platform", "agents", "modules", "api"]

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
exclude = [
    "tests/",
    "infra/",
    "ui/",
    "knowledge_bases/",
    "docs/",
]

[[tool.mypy.overrides]]
module = [
    "qdrant_client.*",
    "fastembed.*",
    "rank_bm25.*",
    "magic.*",
    "spacy.*",
    "rapidfuzz.*",
    "celery.*",
    "networkx.*",
    "langfuse.*",
    "langchain_anthropic.*",
    "langchain_core.*",
    "langgraph.*",
]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["platform", "agents", "modules", "api"]
omit = ["*/tests/*", "*/testing/*"]

[tool.coverage.report]
show_missing = true
fail_under = 0  # raised per-layer as coverage grows
